services/api_gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
asyncpg==0.29.0
pydantic==2.8.2
python-json-logger==2.0.7
nats-py==2.6.0
pyjwt==2.9.0
httpx==0.27.2
